Refactor Connection and Session tests
